+++ src/ch/ehi/ili2db/fromxtf/ClassStat.java
@@ -1,7 +1,7 @@
 package ch.ehi.ili2db.fromxtf;
 
 public class ClassStat{
-	public ClassStat(String tag, int startid) {
+	public ClassStat(String tag, long startid) {
 		super();
 		this.tag = tag;
 		this.objcount = 1;
@@ -9,23 +9,23 @@ public class ClassStat{
 		this.endid = startid;
 	}
 	private String tag=null;
-	private int objcount=0;
-	private int startid=0;
-	private int endid=0;
-	public int getObjcount() {
+	private long objcount=0;
+	private long startid=0;
+	private long endid=0;
+	public long getObjcount() {
 		return objcount;
 	}
-	public int getEndid() {
+	public long getEndid() {
 		return endid;
 	}
-	public void addEndid(int endid) {
+	public void addEndid(long endid) {
 		this.endid = endid;
 		this.objcount++;
 	}
 	public String getTag() {
 		return tag;
 	}
-	public int getStartid() {
+	public long getStartid() {
 		return startid;
 	}
 }
